Copy pasted everything from Jupyter notebook
I copy pasted everything I had from the jupyter notebook into here. A lot of this code is really bad. There are a lot (and I mean a lot) of linting errors.

The whole of the timestamp value imputation can be done away with and replaced with relative location imputation developed on the Kenyan sheets.

There's a lot of work here to do...